refactor: Move TripDialogDesktopContent in a specific component
